+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changes
 =======
 
+Version 1.0.6, 2019/02/26
+-------------------------
+
+  - miner-causalinf-pre
+  - miner-causalinf-post
+
+
 Version 1.0.5, 2019/02/25
 -------------------------